last change

+++ gym_mapf/mapf_env/mapf_env.h
@@ -35,7 +35,7 @@ public:
 
 inline size_t hash<MultiAgentState>::operator()(const MultiAgentState &s) const {
     uint8_t primes[]{2, 3, 5, 7, 11, 13, 17, 19, 23, 29, 31, 37, 41, 43, 47, 53, 59, 61, 67, 71, 73, 79, 83, 89, 97};
-    size_t h = 0;
+    double h = 0;
     int i = 0;
 
     for (i = 0; i < s.locations.size(); i++) {
@@ -103,7 +103,7 @@ public:
 
 size_t hash<MultiAgentAction>::operator()(const MultiAgentAction &a) const {
     uint8_t primes[]{2, 3, 5, 7, 11, 13, 17, 19, 23, 29, 31, 37, 41, 43, 47, 53, 59, 61, 67, 71, 73, 79, 83, 89, 97};
-    size_t h = 0;
+    double h = 0;
     for (size_t i = 0; i < a.actions.size(); ++i) {
         h += pow(primes[i % N_PRIMES], (int) a.actions[i]);
     }
